+++ src/components/index.js
@@ -1,7 +1,7 @@
-export { default as TeacherHeader } from './header/teacherHeader';
-export { default as DirectorHeader } from './header/directorHeader';
+export { default as TeacherHeader } from './layouts/header/teacherHeader';
+export { default as DirectorHeader } from './layouts/header/directorHeader';
 
-export { default as TeacherSideBar } from './sidebar/teacherSidebar';
+export { default as TeacherSideBar } from './layouts/sidebar/teacherSidebar';
 
 export { default as Title } from './textLayout/title';
 export { default as TitleMedium } from './textLayout/titleMedium';
@@ -16,7 +16,7 @@ export { default as SimpleDialog } from './dialog/simpleDialog';
 export { default as Teacher } from './layouts/teacher';
 export { default as Director } from './layouts/director';
 
-export { default as TransferList } from './transferList';
+export { default as TransferList } from './list/transferList';
 
 export { default as Notice } from './notice/notice';
